Separate search function etc. from api.py
